% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.treat.R
\name{df.treat}
\alias{df.treat}
\title{Data Frame Function}
\usage{
df.treat(df)
}
\arguments{
\item{df}{is the dataframe that contains the kuni data}
}
\description{
splits the data into dataframes for each treatment category. Treatments are ecogain, ecoloss, econgain, econloss, and control
}
\author{
Elizabeth Hiroyasu
}

