% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_func.R
\name{support_func}
\alias{support_func}
\title{Support_func}
\usage{
support_func(support.df, treat.type, df)
}
\arguments{
\item{support.df}{is a subset of the kuni data where support.df <- as.numeric(df$project_support==1)}

\item{treat.type}{vector of treatments that you are interested in pooling, note that the third element should always be "control"}

\item{df}{is the dataframe that contains the kuni data or a subset of the kuni data, that indicates the support for 
the management policy. It must contain a column that specifies the treatment category of each observation.}
}
\description{
A function to calculated the mean support, mean control, sample size, standard error, average treatment effect, and 
confidence intervals for a treatment group.
}
\author{
Elizabeth Hiroyasu
}

