% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat_func.R
\name{treat_func}
\alias{treat_func}
\title{individual treatment function}
\usage{
treat_func(df)
}
\arguments{
\item{df}{is the dataframe that contains the kuni data}
}
\description{
splits the data into dataframes for each treatment category then calculates the mean, 
average treatement effect, and standard error.
}
\author{
Elizabeth Hiroyasu
}
