% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_effect.R
\name{treatment_effect}
\alias{treatment_effect}
\title{treatment_effect
Calculating the mean, average treatment effect, and standard errors for a subset of the kuni data}
\usage{
treatment_effect(df)
}
\arguments{
\item{df}{is the dataframe that contains the kuni data or a subset of the kuni data, that indicates the support for 
the management policy. It must contain a column that specifies the treatment category of each observation.}
}
\description{
treatment_effect
Calculating the mean, average treatment effect, and standard errors for a subset of the kuni data
}
\author{
Elizabeth Hiroyasu
}

